partially reverted 824d30f4083731739ce94dde9103e1a1141a267d: Imath::Matrix::operator[] takes an int, not a size_t.

+++ src/appleseed/foundation/math/matrix.h
@@ -1039,9 +1039,9 @@ inline Matrix<T, 3, 3>::Matrix(const Imath::Matrix33<T>& rhs)
 {
     T* p = m_comp;
 
-    for (size_t i = 0; i < 3; ++i)
+    for (int i = 0; i < 3; ++i)
     {
-        for (size_t j = 0; j < 3; ++j)
+        for (int j = 0; j < 3; ++j)
             *p++ = rhs[j][i];
     }
 }
@@ -1564,9 +1564,9 @@ inline Matrix<T, 4, 4>::Matrix(const Imath::Matrix44<T>& rhs)
 {
     T* p = m_comp;
 
-    for (size_t i = 0; i < 4; ++i)
+    for (int i = 0; i < 4; ++i)
     {
-        for (size_t j = 0; j < 4; ++j)
+        for (int j = 0; j < 4; ++j)
             *p++ = rhs[j][i];
     }
 }
